+++ src/main/java/subway/application/StationService.java
@@ -1,6 +1,7 @@
 package subway.application;
 
 import org.springframework.stereotype.Service;
+import subway.application.repository.StationRepository;
 import subway.domain.Line;
 import subway.domain.Section;
 import subway.domain.Sections;
@@ -11,7 +12,6 @@ import subway.dto.StationResponse;
 import subway.entity.LineEntity;
 import subway.entity.SectionEntity;
 import subway.entity.StationEntity;
-import subway.persistence.repository.StationRepository;
 
 import java.util.List;
 import java.util.Optional;
